Multiple Duel types

+++ KitPvP/src/com/minecraft/plugin/elite/kitpvp/listeners/DuelEventListener.java
@@ -5,11 +5,13 @@ import com.minecraft.plugin.elite.general.api.ePlayer;
 import com.minecraft.plugin.elite.kitpvp.KitPvP;
 import com.minecraft.plugin.elite.kitpvp.KitPvPLanguage;
 import com.minecraft.plugin.elite.kitpvp.manager.duel.Duel;
-import com.minecraft.plugin.elite.kitpvp.manager.duel.DuelGUI;
+import com.minecraft.plugin.elite.kitpvp.manager.duel.custom.CustomDuelSelector;
+import com.minecraft.plugin.elite.kitpvp.manager.duel.custom.DuelGUI;
 import com.minecraft.plugin.elite.kitpvp.manager.duel.DuelManager;
 import com.minecraft.plugin.elite.kitpvp.manager.duel.DuelRequest;
 import com.minecraft.plugin.elite.kitpvp.manager.duel.DuelSelectorTool;
-import com.minecraft.plugin.elite.kitpvp.manager.duel.DuelSetup;
+import com.minecraft.plugin.elite.kitpvp.manager.duel.custom.DuelSetup;
+import com.minecraft.plugin.elite.kitpvp.manager.duel.normal.NormalDuelSelectorTool;
 import org.bukkit.Bukkit;
 import org.bukkit.Material;
 import org.bukkit.entity.Player;
@@ -26,44 +28,51 @@ import org.bukkit.inventory.ItemStack;
 public class DuelEventListener implements Listener {
 
 	@EventHandler
-	public void on1v1ItemClick(PlayerInteractEntityEvent e) {
+	public void onDuelSelectorClick(PlayerInteractEntityEvent e) {
 		if(e.getRightClicked() instanceof Player) {
 			ePlayer p = ePlayer.get(e.getPlayer());
 			ePlayer z = ePlayer.get((Player) e.getRightClicked());
 			ItemStack item = p.getPlayer().getItemInHand();
 			if(item.getType() == Material.AIR || item.getType() == null)
 				return;
-			if(!p.isAdminMode() && !p.isWatching() && !z.isAdminMode() && !z.isWatching()) {
+			if(!p.isAdminMode() && !p.isWatching() && !z.isAdminMode() && !z.isWatching() && p.isInRegion(KitPvP.REGION_DUEL) && z.isInRegion(KitPvP.REGION_DUEL)) {
 				if(p.hasTool()) {
-					for(Tool tool : p.getTools())
-						if (item.getItemMeta().hasDisplayName() && tool instanceof DuelSelectorTool) {
-							if (item.getItemMeta().getDisplayName().equalsIgnoreCase(tool.getName())) {
-								Duel duel = DuelManager.get(z);
-								if (duel != null)
-									p.getPlayer().sendMessage(p.getLanguage().get(KitPvPLanguage.DUEL_ALREADY)
-											.replaceAll("%z1", z.getName())
-											.replaceAll("%z2", duel.getOpponent(z).getName()));
-								DuelRequest valid_request = DuelManager.getRequest(z, p);
-								if (valid_request != null) {
-									Duel new_duel = new Duel(valid_request);
-									new_duel.openGUI();
-									return;
+					for(Tool tool : p.getTools()) {
+						if(item.getItemMeta().hasDisplayName()) {
+							String itemName = item.getItemMeta().getDisplayName();
+							if(itemName.equalsIgnoreCase(tool.getName())) {
+								if(tool instanceof CustomDuelSelector) {
+									Duel duel = DuelManager.get(z);
+									if(duel != null)
+										p.getPlayer().sendMessage(p.getLanguage().get(KitPvPLanguage.DUEL_ALREADY)
+												.replaceAll("%z1", z.getName())
+												.replaceAll("%z2", duel.getOpponent(z).getName()));
+									DuelRequest valid_request = DuelManager.getRequest(z, p);
+									if(valid_request != null && valid_request.getType() == Duel.DuelType.CUSTOM) {
+										Duel new_duel = new Duel(valid_request);
+										new_duel.openGUI();
+										return;
+									}
+									if(DuelManager.getRequest(p, z) == null) {
+										DuelRequest reqOld = DuelManager.getRequest(p);
+										if(reqOld != null)
+											reqOld.delete();
+										final DuelRequest request = new DuelRequest(p, z);
+										z.getPlayer().sendMessage(z.getLanguage().get(KitPvPLanguage.DUEL_REQUEST_RECEIVED)
+												.replaceAll("%p", p.getName()));
+										p.getPlayer().sendMessage(p.getLanguage().get(KitPvPLanguage.DUEL_REQUEST_SENT)
+												.replaceAll("%z", z.getName()));
+										Bukkit.getScheduler().scheduleSyncDelayedTask(KitPvP.getPlugin(), request::delete, 200);
+									} else
+										p.sendMessage(KitPvPLanguage.DUEL_REQUEST_COOLDOWN);
 								}
-								if (DuelManager.getRequest(p, z) == null) {
-									DuelRequest reqOld = DuelManager.getRequest(p);
-									if (reqOld != null)
-										reqOld.delete();
-									final DuelRequest request = new DuelRequest(p, z);
-									z.getPlayer().sendMessage(z.getLanguage().get(KitPvPLanguage.DUEL_REQUEST_RECEIVED)
-											.replaceAll("%p", p.getName()));
-									p.getPlayer().sendMessage(p.getLanguage().get(KitPvPLanguage.DUEL_REQUEST_SENT)
-											.replaceAll("%z", z.getName()));
-									Bukkit.getScheduler().scheduleSyncDelayedTask(KitPvP.getPlugin(), request::delete, 200);
-								} else
-									p.sendMessage(KitPvPLanguage.DUEL_REQUEST_COOLDOWN);
+								if(tool instanceof NormalDuelSelectorTool) {
+
+								}
 							}
 							return;
 						}
+					}
 				}
 			}
 		}
